modified forms.py to be useable for edit, updated to semi-class based views

+++ apps/products/models.py
@@ -4,7 +4,7 @@ from django.utils import timezone
 class Product(models.Model):
     name = models.CharField(max_length = 50)
     manufacturer = models.CharField(max_length = 50)
-    price = models.PositiveIntegerField()
+    price = models.DecimalField(max_digits=10,decimal_places=2)
     description = models.CharField(max_length = 50)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
